+++ ts_src/tools/tool.ts
@@ -3,5 +3,7 @@ export abstract class Tool {
     abstract onMouseDown(e: MouseEvent): void;
     abstract onMouseMove(e: MouseEvent): void;
     abstract onMouseUp(e: MouseEvent): void;
+    onSelect() {};
+    onDeselect() {};
     onContextMenu(e: MouseEvent) { };
 }